refactor: LoginForm component (#987)
Refactor the LoginForm component to become a fully functional component.

Instead of having inside knowledge of the login flow, it is now configured via parent props. Instead of hardcoding if-else statements to decide which components to show, it now accepts children as props. It has also been upgraded to become a fully controlled component via the value prop.
